Centralized time bucketing
